refactor component folder into module folder
